pyproject.toml
[tool.poetry]
name = "100x-orchestrator"
version = "0.1.0"
description = "An orchestration system for managing AI coding agents"
authors = ["Aj"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
gitpython = "^3.1.31"
pydantic = "^2.0.2"
pytest = "^7.4.0"
httpx = "^0.24.1"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest-cov = "^4.1.0"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"